+++ packages/cli/services/trialDatastore/grpcservers/datalogServer_test.go
@@ -108,8 +108,21 @@ func TestRunTrialDatalogSimple(t *testing.T) {
 			Msg: &grpcapi.RunTrialDatalogInput_TrialParams{
 				TrialParams: &grpcapi.TrialParams{
 					Actors: []*grpcapi.ActorParams{{
-						Name: "myactor",
-					}},
+						Name:       "myactor",
+						ActorClass: "class1",
+					},
+						{
+							Name:       "myactor2",
+							ActorClass: "class1",
+						},
+						{
+							Name:       "myactor3",
+							ActorClass: "class2",
+						},
+						{
+							Name:       "myactor4",
+							ActorClass: "class2",
+						}},
 					MaxSteps: 72,
 				},
 			},
@@ -137,6 +150,10 @@ func TestRunTrialDatalogSimple(t *testing.T) {
 						ReceiverName: "myactor",
 						Value:        12,
 						TickId:       0,
+						Sources: []*grpcapi.RewardSource{{
+							Value:      12,
+							Confidence: 1,
+						}},
 					}},
 				},
 			},
@@ -159,6 +176,104 @@ func TestRunTrialDatalogSimple(t *testing.T) {
 		assert.Equal(t, float32(12.0), *sample.ActorSamples[0].Reward)
 	}
 	{
+		// Send a sample with multiple rewards for the same actor
+		// Test different combinations with the wildcard character '*'
+		err = stream.Send(&grpcapi.RunTrialDatalogInput{
+			Msg: &grpcapi.RunTrialDatalogInput_Sample{
+				Sample: &grpcapi.DatalogSample{
+					Info: &grpcapi.SampleInfo{
+						TickId: 0,
+					},
+					Rewards: []*grpcapi.Reward{{
+						ReceiverName: "*",
+						Value:        12,
+						TickId:       0,
+						Sources: []*grpcapi.RewardSource{{
+							Value:      12,
+							Confidence: 1,
+						}},
+					},
+						{
+							ReceiverName: "myactor",
+							Value:        8,
+							TickId:       0,
+							Sources: []*grpcapi.RewardSource{{
+								Value:      24,
+								Confidence: 0.5,
+							}},
+						},
+						{
+							ReceiverName: "*.*",
+							Value:        8,
+							TickId:       0,
+							Sources: []*grpcapi.RewardSource{{
+								Value:      24,
+								Confidence: 0.5,
+							}},
+						},
+						{
+							ReceiverName: "class2.*",
+							Value:        8,
+							TickId:       0,
+							Sources: []*grpcapi.RewardSource{{
+								Value:      6,
+								Confidence: 1,
+							}},
+						},
+						{
+							ReceiverName: "class2.myactor4",
+							Value:        8,
+							TickId:       0,
+							Sources: []*grpcapi.RewardSource{{
+								Value:      60,
+								Confidence: 0.5,
+							}},
+						},
+						{
+							ReceiverName: "",
+							Value:        8,
+							TickId:       0,
+							Sources: []*grpcapi.RewardSource{{
+								Value:      24,
+								Confidence: 0.5,
+							}},
+						},
+						{
+							ReceiverName: "class1",
+							Value:        8,
+							TickId:       0,
+							Sources: []*grpcapi.RewardSource{{
+								Value:      24,
+								Confidence: 0.5,
+							}},
+						},
+					},
+				},
+			},
+		})
+
+		assert.NoError(t, err)
+		assert.Equal(t, 2, <-ack)
+	}
+	{
+		// Make sure it is retrieved, and that both rewards are aggregated
+		samples := make(chan *grpcapi.StoredTrialSample)
+		go func() {
+			err := fxt.backend.ObserveSamples(fxt.ctx, backend.TrialSampleFilter{TrialIDs: []string{trialID}}, samples)
+			assert.NoError(t, err)
+		}()
+		<-samples // Skip the first sample
+		sample := <-samples
+		assert.Equal(t, trialID, sample.TrialId)
+		assert.Equal(t, uint64(0), sample.TickId)
+		assert.Nil(t, sample.ActorSamples[0].Action)
+		assert.Nil(t, sample.ActorSamples[0].Observation)
+		assert.Equal(t, float32(18.0), *sample.ActorSamples[0].Reward)
+		assert.Equal(t, float32(16.0), *sample.ActorSamples[1].Reward)
+		assert.Equal(t, float32(12.0), *sample.ActorSamples[2].Reward)
+		assert.Equal(t, float32(20.0), *sample.ActorSamples[3].Reward)
+	}
+	{
 		// Send a sample with an observation and an action
 		err = stream.Send(&grpcapi.RunTrialDatalogInput{
 			Msg: &grpcapi.RunTrialDatalogInput_Sample{
@@ -179,7 +294,7 @@ func TestRunTrialDatalogSimple(t *testing.T) {
 			},
 		})
 		assert.NoError(t, err)
-		assert.Equal(t, 2, <-ack)
+		assert.Equal(t, 3, <-ack)
 	}
 	{
 		// Make sure it is retrieved
@@ -188,7 +303,8 @@ func TestRunTrialDatalogSimple(t *testing.T) {
 			err := fxt.backend.ObserveSamples(fxt.ctx, backend.TrialSampleFilter{TrialIDs: []string{trialID}}, samples)
 			assert.NoError(t, err)
 		}()
-		<-samples // Skip the first sample
+		<-samples // Skip the first two samples
+		<-samples
 		sample := <-samples
 		assert.Equal(t, trialID, sample.TrialId)
 		assert.Equal(t, uint64(1), sample.TickId)